프로젝트 마무리

+++ src/main/java/com/ljy/videoclass/services/panelist/domain/event/SignUpedPanelistEvent.java
@@ -11,7 +11,7 @@ import lombok.Getter;
 public class SignUpedPanelistEvent {
     private String id;
     private String password;
-    private short expellCount;
+    private int expellCount;
     private PanelistStatus status;
 
     @Builder
